CW 130_79

+++ analyzer/src/main/java/telran/probes/AnalyzerAppl.java
@@ -35,7 +35,7 @@ String deviationBindingName;
 		SensorRange range = providerService.getSensorRange(sensorId);
 		float value = probeData.value();
 		
-		int border = 0;
+		float border = 0;
 		if (value < range.minValue()) {
 			border = range.minValue();
 		} else if(value > range.maxValue()) {
